requirements.txt
Django>=3.0,<4.0
gunicorn==20.0.4

psycopg2>=2.8 # PostgreSQL database adapter for Python

pytz==2021.3

whitenoise==5.3.0
django-environ==0.9.0
django-allauth==0.51.0
django-sslserver==0.22

django-crispy-forms==1.14.0
crispy-bootstrap5==0.7
django-import-export==3.0.2

django-admin-honeypot-advanced==1.0.1
django-admin-interface==0.24.2
django-admin-datta==1.0.7